LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-wheel-event-expected.txt

Test for wheel operations for <input type=month>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value is 1999-01. We'll wheel up by 1:
PASS input.value is "1999-02"
Wheel up by 100:
PASS input.value is "1999-03"
Wheel down by 1:
PASS input.value is "1999-02"
Wheel down by 256:
PASS input.value is "1999-01"
Disabled input element:
PASS input.value is "1999-01"
Read-only input element:
PASS input.value is "1999-01"
No focus:
PASS input.value is "1999-01"
PASS successfullyParsed is true

TEST COMPLETE
